Fix. Разделение логики сериализации и конфигурации
