+++ ObjectPrinting/Configurations/Interfaces/ITypePrintingConfig.cs
@@ -6,6 +6,6 @@ namespace ObjectPrinting.Configurations.Interfaces;
 public interface ITypePrintingConfig<TOwner>
 {
     public PrintingConfig<TOwner> ParentConfig { get; }
-    public Func<object, string>? Serializer { get; } 
+    public Func<object, string>? Serializer { get; }
     public CultureInfo? CultureInfo { get; }
 }
