+++ ObjectPrinting/Constants/PrintingConfigConstants.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace ObjectPrinting.Constants;
 
-public static class PrintingConfigConstants
+internal static class PrintingConfigConstants
 {
     public static char Tab => '\t';
     public static string NewLine => Environment.NewLine;
