+++ ObjectPrinting/Extensions/TypePrintingConfigExtensions.cs
@@ -7,7 +7,7 @@ namespace ObjectPrinting.Extensions;
 public static class TypePrintingConfigExtensions
 {
     public static PrintingConfig<TOwner> Using<TOwner, TType>(
-            this TypePrintingConfig<TOwner, TType> typeConfig, CultureInfo culture)
+        this TypePrintingConfig<TOwner, TType> typeConfig, CultureInfo culture)
         where TType : IFormattable
     {
         ArgumentNullException.ThrowIfNull(culture);
